failedCommit renamed to selectedCommit
