+++ src/repairnator-realtime/src/main/java/fr/inria/spirals/repairnator/realtime/githubapi/commits/models/SelectedCommit.java
@@ -64,7 +64,7 @@ public class SelectedCommit {
 
     @Override
     public String toString() {
-        return "FailedCommit{" +
+        return "SelectedCommit{" +
                 "commitId='" + commitId + '\'' +
                 ", repoName='" + repoName + '\'' +
                 '}';
